refactor: changed the data type of price from int to double

+++ src/analytics/ExecutionStats.h
@@ -5,5 +5,5 @@ struct ExecutionStats
     double totalValue = 0;
     int filledQty = 0;
     int intendedQty = 0;
-    int expectedPrice = 0; // this should be double
+    double expectedPrice = 0; // this should be double
 };
